refactor(android): disable react native flipper
