+++ android/app/src/main/java/com/dccms/zanyengo/MainActivity.java
@@ -1,4 +1,4 @@
-package com.weatherapp;
+package com.dccms.zanyengo;
 
 import android.os.Bundle;
 import com.facebook.react.ReactActivity;
